refactor: parentnames

+++ src/Entities/AuditEntities/OriginADT.ts
@@ -10,10 +10,12 @@ export class OriginADT extends Origin implements AuditingEntityDefaultColumns {
     readonly _modifiedAt!: Date;
 
     @BeforeInsert()
-    dropFkParentId() {
-        const queryRunner = appDataSource.createQueryRunner();
-        queryRunner.query('ALTER TABLE adt_origin DROP COLUMN fk_definition_id;');
-        queryRunner.query('ALTER TABLE adt_origin ADD COLUMN fk_definition_id INT NULL;');
-        queryRunner.release();
+    dropFkOriginId() {
+        if (process.env.SYNCHRONIZE === 'true') {
+            const queryRunner = appDataSource.createQueryRunner();
+            queryRunner.query('ALTER TABLE adt_origin DROP COLUMN fk_definition_id;');
+            queryRunner.query('ALTER TABLE adt_origin ADD COLUMN fk_definition_id INT NULL;');
+            queryRunner.release();
+        }
     }
 }
